# public/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makkelijke Markt API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/1.1.0/sollicitaties/markt/{marktId}": {
            "get": {
                "tags": [
                    "Sollicitatie"
                ],
                "summary": "Vraag sollicitaties op voor een markt",
                "operationId": "SollicitatieGetAllByMarktIdWithFilter",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeDoorgehaald",
                        "in": "query",
                        "description": "Default=1",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "listOffset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "listLength",
                        "in": "query",
                        "description": "Default=100",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Sollicitatie"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/sollicitaties/id/{id}": {
            "get": {
                "tags": [
                    "Sollicitatie"
                ],
                "summary": "Gegevens van sollicitatie op basis van id",
                "operationId": "SollicitatiesGetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sollicitatie"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/sollicitaties/markt/{marktId}/{sollicitatieNummer}": {
            "get": {
                "tags": [
                    "Sollicitatie"
                ],
                "summary": "Gegevens van sollicitatie op basis van markt en sollicitatienummer",
                "operationId": "SollicitatiesGetAllByMarktAndSollicitatieNummer",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sollicitatieNummer",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sollicitatie"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/lijst/week/{marktId}": {
            "get": {
                "tags": [
                    "Lijst"
                ],
                "summary": "Weeklijst voor markt",
                "operationId": "SollicitatieGetAllPerWeekByMarktId",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Sollicitatie"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/lijst/week/{marktId}/{types}": {
            "get": {
                "tags": [
                    "Lijst"
                ],
                "summary": "Weeklijst voor markt op basis van sollicatie types",
                "operationId": "SollicitatieGetAllPerWeekByMarktIdAndTypes",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "types",
                        "in": "path",
                        "description": "Koopman types gescheiden met een | zoals: soll, vpl, vkk",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Sollicitatie"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/lijst/week/{marktId}/{types}/{startDate}/{endDate}": {
            "get": {
                "tags": [
                    "Lijst"
                ],
                "summary": "Weeklijst voor markt op basis van sollicatie types en datum",
                "operationId": "SollicitatieGetAllPerWeekByMarktIdAndTypesAndDates",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "types",
                        "in": "path",
                        "description": "Koopman types gescheiden met een | zoals: soll, vpl, vkk",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "path",
                        "description": "date as yyyy-mm-dd",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "description": "date as yyyy-mm-dd",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Sollicitatie"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Koopman": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "erkenningsnummer": {
                        "type": "string"
                    },
                    "voorletters": {
                        "type": "string"
                    },
                    "tussenvoegsels": {},
                    "achternaam": {
                        "type": "string"
                    },
                    "email": {},
                    "telefoon": {},
                    "status": {
                        "type": "integer"
                    },
                    "koopmanStatus": {},
                    "fotoUrl": {},
                    "fotoMediumUrl": {},
                    "pasUid": {},
                    "perfectViewNummer": {},
                    "handhavingsVerzoek": {},
                    "weging": {}
                },
                "type": "object"
            },
...
            "Markt": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "afkorting": {
                        "type": "string",
                        "example": "NOM-M"
                    },
                    "naam": {
                        "type": "string",
                        "example": "Noordermarkt maandag"
                    },
                    "soort": {
                        "type": "string",
                        "example": "week"
                    },
                    "geoArea": {},
                    "marktDagen": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "ma",
                            "di"
                        ]
                    },
                    "standaardKraamAfmeting": {
                        "example": 3
                    },
                    "extraMetersMogelijk": {},
                    "aanwezigeOptiesResult": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": {
                            "3mKramen": true,
                            "4mKramen": false,
                            "extraMeters": true,
                            "elektra": true
                        }
                    },
                    "perfectViewNummer": {
                        "type": "integer",
                        "example": 40
                    },
                    "aantalKramen": {
                        "type": "integer"
                    },
                    "aantalMeter": {
                        "type": "integer"
                    },
                    "auditMax": {
                        "type": "integer",
                        "example": 10
                    },
                    "kiesJeKraamMededelingActief": {
                        "type": "boolean"
                    },
                    "kiesJeKraamMededelingTitel": {
                        "type": "string"
                    },
                    "kiesJeKraamMededelingTekst": {
                        "type": "string"
                    },
                    "kiesJeKraamActief": {
                        "type": "boolean"
                    },
                    "kiesJeKraamFase": {
                        "type": "string"
                    },
                    "kiesJeKraamGeblokkeerdePlaatsen": {
                        "type": "string"
                    },
                    "kiesJeKraamGeblokkeerdeData": {
                        "type": "string",
                        "example": "2019-12-25,2019-12-26"
                    },
                    "kiesJeKraamEmailKramenzetter": {
                        "type": "string",
                        "example": "test@domain.tld"
                    },
                    "marktDagenTekst": {
                        "type": "string"
                    },
                    "indelingsTijdstipTekst": {
                        "type": "string"
                    },
                    "telefoonNummerContact": {
                        "type": "string"
                    },
                    "makkelijkeMarktActief": {
                        "type": "boolean"
                    },
                    "indelingstype": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Sollicitatie": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "sollicitatieNummer": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "vastePlaatsen": {},
                    "aantal3MeterKramen": {},
                    "aantal4MeterKramen": {},
                    "aantalExtraMeters": {},
                    "aantalElektra": {},
                    "aantalAfvaleilanden": {
                        "type": "integer"
                    },
                    "krachtstroom": {},
                    "doorgehaald": {
                        "type": "boolean"
                    },
                    "doorgehaaldReden": {},
                    "koppelveld": {},
                    "markt": {
                        "$ref": "#/components/schemas/Markt"
                    },
                    "koopman": {
                        "$ref": "#/components/schemas/Koopman"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "MmAppKey",
                "in": "header"
            },
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}